Moved AT-RT config
